kloud: move build method to it's own file
